refactor decode runner
